+++ src/Symfony/Component/Validator/Constraints/CountValidator.php
@@ -60,6 +60,20 @@ class CountValidator extends ConstraintValidator
                 ->setPlural((int) $constraint->min)
                 ->setCode(Count::TOO_FEW_ERROR)
                 ->addViolation();
+
+            return;
+        }
+
+        if (null !== $constraint->divisibleBy) {
+            $this->context
+                ->getValidator()
+                ->inContext($this->context)
+                ->validate($count, [
+                    new DivisibleBy([
+                        'value' => $constraint->divisibleBy,
+                        'message' => $constraint->divisibleByMessage,
+                    ]),
+                ], $this->context->getGroup());
         }
     }
 }